# openapi.yaml
openapi: 3.1.0
info:
  title: QuizForge API
  version: 0.5.0-dev
  summary: Asynchronous quiz and assessment tournaments.
  description: |
    QuizForge runs asynchronous quiz tournaments. A tournament opens for a
    window, players enter on their own schedule, and results roll up to a
    leaderboard. There is no synchronous "everyone answers at once" mode, by
    design.

    ## Status of this document

    This describes the API **exactly as it behaves today**.

    The shapes it first described were not all ones worth keeping. Six were
    corrected before anything consumed this document — while doing so was one
    commit rather than a breaking change — and are recorded in
    `docs/api/known-inconsistencies.md` with what they were and why they moved.

    ## Conventions

    - Every identifier is a prefixed `TypeId`: `acc_`, `wsp_`, `trn_`, `att_`,
      `qst_`. Bare UUIDs never appear in a response.
    - Every list endpoint returns `{ "data": [...], "nextCursor": ... }` and
      takes `limit` and `cursor`. Pagination is **keyset**, never offset: a
      cursor names the last row you saw, so a row inserted while you page
      through cannot duplicate or hide another one.
    - Out-of-range parameters are **rejected**, never silently corrected.

    Everything the product does is reachable here: workspaces, members, API
    keys, question banks, questions, tournaments and play.

    ## Questions are immutable

...
    question.

...

...

...

...
    have selected.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - bearerApiKey: []
paths:
...
  /v1/members:
    get:
      tags: [Members]
      operationId: listMembers
      summary: Who belongs to this workspace
      description: |
        Every member of the workspace in scope, with their role.

        Carries each member's email, so this is readable by any member rather
        than only by administrators — a `VIEWER` who can see the leaderboard can
        already see who is playing.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: The workspace's members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '400':
          $ref: '#/components/responses/InvalidListRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '500':
          $ref: '#/components/responses/Internal'

    post:
      tags: [Members]
      operationId: addMember
      summary: Add an existing account to this workspace
      description: |
        Adds an account, identified by email, with the given role. Requires
        `ADMIN` or `OWNER`.

        **The account must already exist.** There is no invitation flow yet, so
        an unregistered address is `404` rather than a pending invite.

        That `404` is an account-enumeration signal, and is documented rather
        than hidden: an administrator of any workspace can learn whether an
        address has an account. The exposure is narrower than it looks — the
        caller must already hold a privileged role somewhere — but it is real
        and disappears when invitations land.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: The account is now a member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          description: |
            No account has that email address. Carries `code: NOT_FOUND`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: |
            That account is already a member. Carries `code: ALREADY_EXISTS`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/members/{accountId}:
    patch:
      tags: [Members]
      operationId: changeMemberRole
      summary: Change a member's role
      description: |
        Requires `ADMIN` or `OWNER`.

        Demoting the **last** `OWNER` is refused with `400`. A workspace with no
        owner cannot issue keys, manage members or be renamed by anybody, and
        nothing in the API could undo it.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRoleRequest'
      responses:
        '200':
          description: The role was changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: |
            The role is invalid, or this would demote the last `OWNER`. Carries
            `code: INVALID_REQUEST`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'

    delete:
      tags: [Members]
      operationId: removeMember
      summary: Remove a member from this workspace
      description: |
        Requires `ADMIN` or `OWNER`. Removing the last `OWNER` is refused with
        `400`, for the same reason demoting one is.

        The account itself is untouched — it keeps its login and any other
        workspaces it belongs to.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '204':
          description: The member was removed.
        '400':
          description: |
            This would remove the last `OWNER`. Carries `code: INVALID_REQUEST`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/api-keys:
    get:
      tags: [API keys]
      operationId: listApiKeys
      summary: The active API keys of this workspace
      description: |
        Metadata only. **The secret is never returned here**, and there is no
        endpoint that returns it again — it exists in a response exactly once,
        when the key is created.

        Revoked keys are omitted rather than listed as revoked. Requires
        `OWNER`.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: The workspace's active keys.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyPage'
        '400':
          $ref: '#/components/responses/InvalidListRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '500':
          $ref: '#/components/responses/Internal'

    post:
      tags: [API keys]
      operationId: createApiKey
      summary: Mint an API key
      description: |
        Creates a key and returns its secret **once**. Only a SHA-256 digest is
        stored, so the plaintext cannot be recovered, re-sent, or read out of a
        database dump. A client that does not persist it must mint another.

        Requires `OWNER`.

        The `environment` decides the prefix: `qf_live_…` or `qf_test_…`. Both
        currently reach the same data — sandbox isolation arrives later — so a
        `test` key today is a labelling convention, not a safety boundary. It is
        offered now so that keys issued today keep working when it becomes one.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: |
            The key was created. This is the only response that will ever
            contain `secret`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedApiKey'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/api-keys/{keyId}:
    delete:
      tags: [API keys]
      operationId: revokeApiKey
      summary: Revoke an API key
      description: |
        Revokes the key immediately. The **next** request presenting it is
        rejected — keys are resolved against the database on every request, so
        there is no cache to expire and no window in which a revoked key still
        works.

        Requires `OWNER`. A key belonging to another workspace is `404`, not
        `403`: its existence elsewhere is not disclosed.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: keyId
          in: path
          required: true
          description: The key to revoke.
          schema:
            $ref: '#/components/schemas/ApiKeyId'
      responses:
        '204':
          description: The key is revoked.
        '400':
          $ref: '#/components/responses/WorkspaceRequired'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qf_session
      description: |
        Opaque session token issued by `POST /v1/auth/login`, stored only as a
        SHA-256 digest on the server. Mutating requests authenticated this way
        are CSRF-protected and additionally require the `X-XSRF-TOKEN` header.
    bearerApiKey:
      type: http
      scheme: bearer
      description: |
        A workspace API key, sent as `Authorization: Bearer qf_…`. Shown once
        when created and stored only as a SHA-256 digest thereafter.

        Authenticates a workspace rather than a person, so endpoints that
        attribute an action to an account — starting and submitting attempts —
        reject it.
  parameters:
    WorkspaceHeader:
      name: X-QuizForge-Workspace
      in: header
      required: false
      description: |
        The workspace to operate on, as a `wsp_…` identifier.

        Required for every endpoint outside `/v1/auth` when authenticating with
        a session cookie; omitting it fails with `400 INVALID_REQUEST`. With an
        API key the workspace comes from the key and this header is ignored.

        Documented as `required: false` because it is genuinely optional for
        key-authenticated callers, not because it is safe to leave out.

        Naming a workspace you are not a member of returns
        **`403 PERMISSION_DENIED`**: the credential is valid, the workspace is
        not yours. A malformed value is `400 INVALID_REQUEST`.
      schema:
        $ref: '#/components/schemas/WorkspaceId'
    Limit:
      name: limit
      in: query
      required: false
      description: |
        How many items to return, 1–100.

        A value outside that range is `400 INVALID_REQUEST`. It is deliberately
        not clamped: answering a request for 1000 with 100 items is
        indistinguishable from a collection that only holds 100, so a client
        would conclude it had everything when it had a prefix.
      schema:
        type: integer
        format: int32
        default: 25
        minimum: 1
        maximum: 100
    PageCursor:
      name: cursor
      in: query
      required: false
      description: |
        The `nextCursor` from the previous page. Omit for the first page.

        **Treat this as opaque.** It is not a page number, an offset, or an
        identifier, and its encoding is not part of the contract. A value this
        API did not issue is `400 INVALID_CURSOR` rather than being quietly
        treated as "start again".
      schema:
        $ref: '#/components/schemas/CursorParameter'
    AccountIdPath:
      name: accountId
      in: path
      required: true
      description: The account whose membership is being changed.
      schema:
        $ref: '#/components/schemas/AccountId'
...
  responses:
    InvalidRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspaceRequired:
      description: |
        No workspace is in scope. Carries `code: INVALID_REQUEST`.

        Sent when a session-authenticated request omits
        `X-QuizForge-Workspace`. Tenant isolation depends on a workspace being
        selected, so such a request is refused before it reaches a handler
        rather than being served unfiltered.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspaceForbidden:
      description: |
        The credential is valid but the account is not a member of the workspace
        it named. Carries `code: PERMISSION_DENIED`.

        Distinct from `401` deliberately: re-authenticating cannot fix this, so
        telling the caller they are unauthenticated would send them to the wrong
        remedy.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationRequired:
      description: |
        No valid credential was presented. Carries
        `code: AUTHENTICATION_REQUIRED`.

        Also returned in two cases that are not obviously authentication
        failures:

        - A path that does not exist. Authentication is evaluated before
          routing, so an anonymous caller learns nothing about which paths are
          real.
        It is **not** returned for a workspace you are not a member of — that
        is `403`, because the credential itself is valid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InvalidListRequest:
      description: |
        The request could not be paged. One of:

        - no workspace is in scope — `code: INVALID_REQUEST`
        - `limit` is outside 1–100 — `code: INVALID_REQUEST`
        - `cursor` was not issued by this API — `code: INVALID_CURSOR`

        The three are distinguished by `code`, which is why they share a status
        without sharing a meaning.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: |
        No such object, **or** it belongs to another workspace or another
        account. The two are deliberately indistinguishable, so this endpoint
        cannot be used to discover that an identifier exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Internal:
      description: |
        Something failed unexpectedly. The cause is logged server-side and
        never included in the response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      title: Problem
      description: |
        An RFC 9457 Problem Details document. Every error in the API has this
        shape.

        Clients should branch on `code`. It is the only field with a stability
        guarantee: a code that exists will never change meaning or HTTP status,
        and new codes are added rather than existing ones repurposed.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          maxLength: 200
          description: Stable URI identifying the error class.
          examples: ['https://quizforge.dev/errors/not-found']
        title:
          type: string
          maxLength: 100
          description: The error code, repeated as a human-readable title.
          examples: ['NOT_FOUND']
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code, repeated in the body.
          examples: [404]
        detail:
          type: string
          maxLength: 2000
          description: |
            Human-readable explanation. **Not stable** — reworded freely. Never
            parse it.
          examples: ['no such attempt']
        instance:
          type: string
          format: uri-reference
          maxLength: 2000
          description: The path that produced the error.
          examples: ['/v1/attempts/att_01a0136fb3937209a5719e8fb7db26bf']
        code:
          $ref: '#/components/schemas/ErrorCode'
...
    AccountId:
      type: string
      title: AccountId
      description: An account identifier. UUIDv7 rendered as 32 lowercase hex characters with an `acc_` prefix.
      pattern: '^acc_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['acc_01a0136fb3937209a5719e8fb7db26bf']
    WorkspaceId:
      type: string
      title: WorkspaceId
      description: A workspace identifier. UUIDv7 rendered as 32 lowercase hex characters with a `wsp_` prefix.
      pattern: '^wsp_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['wsp_01a0136fb3937209a5719e8fb7db26bf']
...
    Cursor:
      type: [string, 'null']
      title: Cursor
      description: |
        Pass this back as `cursor` to get the next page. `null` means there is
        no next page — stop, rather than asking for one you already know is
        empty.

        **Opaque.** Not a page number, not an offset, not an identifier. Its
        encoding is not part of the contract and may change; a cursor issued by
        an older version will be rejected rather than misread.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
    CursorParameter:
      type: string
      title: CursorParameter
      description: An opaque cursor previously returned as `nextCursor`.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
...
    ApiKeyId:
      type: string
      title: ApiKeyId
      description: An API key identifier. UUIDv7 rendered as 32 lowercase hex characters with a `key_` prefix.
      pattern: '^key_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['key_01a0136fb3937209a5719e8fb7db26bf']
    Role:
      type: string
      title: Role
      description: |
        What a member may do in a workspace, from most to least privileged.

        - `OWNER` — everything, including managing API keys and renaming the
          workspace. A workspace always has at least one.
        - `ADMIN` — manage members, content and tournaments.
        - `EDITOR` — manage content and tournaments.
        - `VIEWER` — read only.
      enum: [OWNER, ADMIN, EDITOR, VIEWER]
      examples: ['OWNER']
    KeyEnvironment:
      type: string
      title: KeyEnvironment
      description: |
        Decides the key's prefix: `qf_live_…` or `qf_test_…`.

        Both currently reach the same data. Sandbox isolation arrives later; the
        distinction exists now so keys issued today keep working when it becomes
        a real boundary.
      enum: [live, test]
      default: live
      examples: ['live']
...
    Member:
      type: object
      title: Member
      description: An account's membership of a workspace.
      required: [accountId, email, displayName, role]
      properties:
        accountId:
          $ref: '#/components/schemas/AccountId'
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          maxLength: 120
        role:
          $ref: '#/components/schemas/Role'
      examples:
        - accountId: acc_01a0136fb3937209a5719e8fb7db26bf
          email: player@example.com
          displayName: Alex Doe
          role: EDITOR
    MemberPage:
      type: object
      title: MemberPage
      required: [data]
      properties:
        data:
          type: array
          maxItems: 10000
          items:
            $ref: '#/components/schemas/Member'
        nextCursor:
          $ref: '#/components/schemas/Cursor'
...
    ApiKeyPage:
      type: object
      title: ApiKeyPage
      required: [data]
      properties:
        data:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/ApiKeySummary'
        nextCursor:
          $ref: '#/components/schemas/Cursor'
    IssuedApiKey:
      type: object
      title: IssuedApiKey
      description: |
        A newly created key, **including its secret**.

        This is the only place `secret` ever appears. Only a SHA-256 digest is
        stored, so it cannot be recovered afterwards by any endpoint, support
        request, or database dump. Store it now or mint another.
      required: [id, name, environment, lastFour, createdAt, secret]
      properties:
        id:
          $ref: '#/components/schemas/ApiKeyId'
        name:
          type: string
          maxLength: 120
        environment:
          $ref: '#/components/schemas/KeyEnvironment'
        lastFour:
          type: string
          minLength: 4
          maxLength: 4
        createdAt:
          type: string
          format: date-time
          maxLength: 40
        secret:
          type: string
          minLength: 8
          maxLength: 200
          description: |
            Send as `Authorization: Bearer <secret>`. Shown once and never
            again.

            A real value is `qf_live_` or `qf_test_` followed by high-entropy
            random characters. The example below is deliberately not of that
            shape: a documentation sample that looks like a live credential
            trips secret scanners, and worse, teaches readers to skim past the
            one that is real.
      examples:
        - id: key_01a0136fb3937209a5719e8fb7db26bf
          name: CI pipeline
          environment: live
          lastFour: 'a9f3'
          createdAt: '2026-08-18T10:00:00Z'
          secret: qf_live_example_value_not_a_real_key
...
    AddMemberRequest:
      type: object
      title: AddMemberRequest
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
          description: Must belong to an existing account; there is no invitation flow yet.
        role:
          $ref: '#/components/schemas/Role'
      examples:
        - email: player@example.com
          role: EDITOR
    ChangeRoleRequest:
      type: object
      title: ChangeRoleRequest
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/Role'
      examples:
        - role: ADMIN
    CreateApiKeyRequest:
      type: object
      title: CreateApiKeyRequest
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: A label, so keys can be told apart when it is time to revoke one.
        environment:
          $ref: '#/components/schemas/KeyEnvironment'
      examples:
        - name: CI pipeline
          environment: live
...